BUGFIX: Zeniths plotted opposite to how they should have been.
The zeniths used to be plotted just as they came out of the data - ie. plotting from 0 to 90, with 0 at the centre and 90 at the edge. This is the WRONG WAY AROUND, and so they have been reversed now by the line:

zenith = 90 - zenith

This now means that the plots are correct, and has a nice side-effect of lining up the plot correctly.